refactored core
