+++ src/commands/backtrace.rs
@@ -1,6 +1,8 @@
 use crate::commands::DebugCommand;
 use crate::core::Debugger;
 use anyhow::Result;
+use crate::core::backtrace::Backtrace;
+
 #[derive(Clone)]
 pub struct BacktraceCommand;
