+++ src/commands/control.rs
@@ -1,6 +1,8 @@
 use crate::commands::DebugCommand;
 use crate::core::Debugger;
 use anyhow::Result;
+use crate::core::stepping::Stepping;
+
 #[derive(Clone)]
 pub struct SingleStepCommand;
