+++ src/commands/registers.rs
@@ -1,6 +1,8 @@
 use crate::commands::DebugCommand;
 use crate::core::Debugger;
 use anyhow::Result;
+use crate::core::registers::Registers;
+
 #[derive(Clone)]
 pub struct SetRegisterCommand;
