+++ src/core/mod.rs
@@ -1,12 +1,13 @@
-mod stepping;
-mod disassembler;
-mod breakpoint;
-mod process_control;
-mod registers;
-mod memory;
-mod backtrace;
-mod debugger;
+pub mod backtrace;
+pub mod breakpoint;
+pub mod debugger;
+pub mod disassembler;
+pub mod memory;
+pub mod process_control;
+pub mod registers;
+pub mod stepping;
+pub mod process;
+pub mod symbols;
+pub mod map;
 
-use stepping::Stepping;
-use disassembler::Disassembler;
-use debugger::*;
+pub use debugger::*;
