+++ src/core/process_control.rs
@@ -1,7 +1,7 @@
-use crate::core::Debugger;
-use anyhow::Result;
+use crate::core::*;
 use log::info;
 use nix::sys::wait::{waitpid, WaitStatus};
+use nix::sys::ptrace::*;
 
 pub trait ProcessControl {
     fn resume_and_wait(&mut self);
