+++ src/main.rs
@@ -1,13 +1,8 @@
 mod commands;
 mod core;
-mod dwarf;
-mod functions;
-mod map;
-mod process;
 mod repl;
-mod stack_unwind;
-use repl::Repl;
 
+use repl::Repl;
 use std::env;
 
 fn main() {
